my_base/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
qiskit>=0.44.0
qiskit-aer>=0.12.0
pennylane>=0.31.0
pennylane-qiskit>=0.31.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
wandb>=0.15.0
tqdm>=4.65.0
pyyaml>=6.0.0
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0
python-dotenv>=1.0.0 